refactor: 스타일 파일 경로 정리
