+++ app/styles/globals.css
@@ -1,5 +1,5 @@
 @import "tailwindcss";
-@import  "../src/styles/theme.css";
+@import  "./theme.css";
 @import "tw-animate-css";
 @import "shadcn/tailwind.css";
