+++ src/shared/lib/cn/cn.ts
@@ -1,6 +1,6 @@
 import { clsx, type ClassValue } from "clsx"
 import { twMerge } from "tailwind-merge"
 
-export function cn(...inputs: ClassValue[]) {
+export const cn=(...inputs: ClassValue[]) =>{
   return twMerge(clsx(inputs))
 }
