+++ src/shared/lib/cn/index.ts
@@ -1,0 +1,1 @@
+export {cn} from './cn'
